Final submission

+++ school/holidays/views.py
@@ -2,12 +2,19 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import Holiday
 from .forms import HolidayForm
 import json
+from django.contrib.auth.decorators import login_required
 
+@login_required
 def holiday_list(request):
+    if getattr(request.user, 'is_student', False):
+        return render(request, 'Home/error_403.html')
     holidays = Holiday.objects.all()
     return render(request, 'holidays/holidays.html', {'holidays': holidays})
 
+@login_required
 def add_holiday(request):
+    if getattr(request.user, 'is_student', False):
+        return render(request, 'Home/error_403.html')
     if request.method == "POST":
         form = HolidayForm(request.POST)
         if form.is_valid():
@@ -18,7 +25,10 @@ def add_holiday(request):
     return render(request, 'holidays/add-holiday.html', {'form': form})
 
 
+@login_required
 def edit_holiday(request, holiday_id):
+    if getattr(request.user, 'is_student', False):
+        return render(request, 'Home/error_403.html')
     holiday = get_object_or_404(Holiday, holiday_id=holiday_id)
     if request.method == 'POST':
         form = HolidayForm(request.POST, instance=holiday)
@@ -29,7 +39,10 @@ def edit_holiday(request, holiday_id):
         form = HolidayForm(instance=holiday)
     return render(request, 'holidays/edit-holiday.html', {'form': form, 'holiday': holiday})
 
+@login_required
 def delete_holiday(request, holiday_id):
+    if getattr(request.user, 'is_student', False):
+        return render(request, 'Home/error_403.html')
     holiday = get_object_or_404(Holiday, holiday_id=holiday_id)
     holiday.delete()
     return redirect('holiday_list')
